% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\docType{data}
\name{credentials}
\alias{credentials}
\title{#' Include Content From a File
#' 
#' Load rendered RMarkdown from a file and turn into HTML.
#' 
#' @rdname includeRMarkdown
#' @export
#' 
#' @importFrom rmarkdown render
#' @importFrom markdown markdownToHTML
#' @importFrom htmltools HTML
includeRMarkdown <- function(path){
  
  md <- tempfile(fileext = '.md')
  
  on.exit(unlink(md),add = TRUE)
  
  rmarkdown::render(
    path,
    output_format = 'md_document',
    output_dir = tempdir(),
    output_file = md,quiet = TRUE
    )
  
  html <- markdown::markdownToHTML(md, fragment.only = TRUE)
  
  Encoding(html) <- "UTF-8"
  
  return(HTML(html))
}}
\format{
An object of class \code{data.frame} with 2 rows and 4 columns.
}
\usage{
credentials
}
\description{
#' Include Content From a File
#' 
#' Load rendered RMarkdown from a file and turn into HTML.
#' 
#' @rdname includeRMarkdown
#' @export
#' 
#' @importFrom rmarkdown render
#' @importFrom markdown markdownToHTML
#' @importFrom htmltools HTML
includeRMarkdown <- function(path){
  
  md <- tempfile(fileext = '.md')
  
  on.exit(unlink(md),add = TRUE)
  
  rmarkdown::render(
    path,
    output_format = 'md_document',
    output_dir = tempdir(),
    output_file = md,quiet = TRUE
    )
  
  html <- markdown::markdownToHTML(md, fragment.only = TRUE)
  
  Encoding(html) <- "UTF-8"
  
  return(HTML(html))
}
}
\keyword{datasets}
